refactor: split video processing tasks
